TextMining_Hands-on/TextMining_Hands-on/newsarticles_streamlit_app/requirements.txt
matplotlib==3.7.0
newspaper3k==0.2.8
nltk==3.7
pandas==1.5.3
streamlit==1.23.1
wordcloud==1.9.2
